+++ crates/bcr-ebill-web/Cargo.toml
@@ -17,12 +17,20 @@ anyhow.workspace = true
 thiserror.workspace = true
 rocket = "0.5.1"
 rocket_cors = { version = "0.6.0", default-features = false }
-rocket_ws = { version = "0.1.1", features = ["tokio-tungstenite"], default-features = false }
+rocket_ws = { version = "0.1.1", features = [
+  "tokio-tungstenite",
+], default-features = false }
 open = "5.3.2"
 clap = { version = "4.5.29", features = ["derive", "env"] }
-utoipa = { version = "5.3.1", features = ["chrono", "rocket_extras"], default-features = false }
-utoipa-swagger-ui = { version = "9.0.0", features = ["rocket"], default-features = false }
+utoipa = { version = "5.3.1", features = [
+  "chrono",
+  "rocket_extras",
+], default-features = false }
+utoipa-swagger-ui = { version = "9.0.0", features = [
+  "rocket",
+], default-features = false }
 bcr-ebill-api = { path = "../bcr-ebill-api" }
+bcr-ebill-transport = { path = "../bcr-ebill-transport" }
 
 [features]
 embedded-db = ["bcr-ebill-api/embedded-db"]
