+++ crates/bcr-ebill-web/src/error.rs
@@ -18,5 +18,5 @@ pub enum Error {
 
     /// all errors originating from the bcr API notification service layer
     #[error("Bill Service error: {0}")]
-    NotificationService(#[from] service::notification_service::Error),
+    NotificationService(#[from] bcr_ebill_transport::Error),
 }
